move function to utils
move CalculateArithmeticShift to utils
